VoteTxDetails component testing
